feat: update creditProfileInfo mock to use new data structure
