+++ src/mocks/creditProfileInfo/creditProfileInfo.mock.ts
@@ -1,0 +1,74 @@
+export const credit_profileInfo = [
+  {
+    credit_request_id: "999990",
+    labor_stability: {
+      company_seniority: "5",
+      labor_stability_index: 900,
+      max_labor_stability_index: 1000,
+      estimated_severance: 20000000,
+    },
+  },
+  {
+    credit_request_id: "999991",
+    labor_stability: {
+      company_seniority: "4",
+      labor_stability_index: 900,
+      max_labor_stability_index: 1000,
+      estimated_severance: 20000000,
+    },
+  },
+  {
+    credit_request_id: "999992",
+    labor_stability: {
+      company_seniority: "15",
+      labor_stability_index: 900,
+      max_labor_stability_index: 1000,
+      estimated_severance: 20000000,
+    },
+  },
+  {
+    credit_request_id: "999993",
+    labor_stability: {
+      company_seniority: "9",
+      labor_stability_index: 900,
+      max_labor_stability_index: 1000,
+      estimated_severance: 20000000,
+    },
+  },
+  {
+    credit_request_id: "999994",
+    labor_stability: {
+      company_seniority: "5",
+      labor_stability_index: 900,
+      max_labor_stability_index: 1000,
+      estimated_severance: 20000000,
+    },
+  },
+  {
+    credit_request_id: "999995",
+    labor_stability: {
+      company_seniority: "3",
+      labor_stability_index: 900,
+      max_labor_stability_index: 1000,
+      estimated_severance: 20000000,
+    },
+  },
+  {
+    credit_request_id: "999996",
+    labor_stability: {
+      company_seniority: "7",
+      labor_stability_index: 900,
+      max_labor_stability_index: 1000,
+      estimated_severance: 20000000,
+    },
+  },
+  {
+    credit_request_id: "999997",
+    labor_stability: {
+      company_seniority: "7",
+      labor_stability_index: 900,
+      max_labor_stability_index: 1000,
+      estimated_severance: 20000000,
+    },
+  },
+];
