--- src/mocks/creditProfileInfo/jobStability.mock.ts
@@ -1,90 +1,0 @@
-export const credit_profileInfo = [
-  {
-    credit_request_id: "999990",
-    job_stability: {
-      labor_stability: {
-        company_seniority: "5",
-        labor_stability_index: 900,
-        max_labor_stability_index: 1000,
-        estimated_severance: 20000000,
-      },
-    },
-  },
-  {
-    credit_request_id: "999991",
-    job_stability: {
-      labor_stability: {
-        company_seniority: "4",
-        labor_stability_index: 900,
-        max_labor_stability_index: 1000,
-        estimated_severance: 20000000,
-      },
-    },
-  },
-  {
-    credit_request_id: "999992",
-    job_stability: {
-      labor_stability: {
-        company_seniority: "15",
-        labor_stability_index: 900,
-        max_labor_stability_index: 1000,
-        estimated_severance: 20000000,
-      },
-    },
-  },
-  {
-    credit_request_id: "999993",
-    job_stability: {
-      labor_stability: {
-        company_seniority: "9",
-        labor_stability_index: 900,
-        max_labor_stability_index: 1000,
-        estimated_severance: 20000000,
-      },
-    },
-  },
-  {
-    credit_request_id: "999994",
-    job_stability: {
-      labor_stability: {
-        company_seniority: "5",
-        labor_stability_index: 900,
-        max_labor_stability_index: 1000,
-        estimated_severance: 20000000,
-      },
-    },
-  },
-  {
-    credit_request_id: "999995",
-    job_stability: {
-      labor_stability: {
-        company_seniority: "3",
-        labor_stability_index: 900,
-        max_labor_stability_index: 1000,
-        estimated_severance: 20000000,
-      },
-    },
-  },
-  {
-    credit_request_id: "999996",
-    job_stability: {
-      labor_stability: {
-        company_seniority: "7",
-        labor_stability_index: 900,
-        max_labor_stability_index: 1000,
-        estimated_severance: 20000000,
-      },
-    },
-  },
-  {
-    credit_request_id: "999997",
-    job_stability: {
-      labor_stability: {
-        company_seniority: "7",
-        labor_stability_index: 900,
-        max_labor_stability_index: 1000,
-        estimated_severance: 20000000,
-      },
-    },
-  },
-];
