+++ src/mocks/utils/initializeDataDB.ts
@@ -11,7 +11,7 @@ import { approval_by_credit_request_Mock } from "@src/mocks/financialReporting/A
 import { documents } from "@mocks/financialReporting/documents.mock";
 import { traceMock } from "@mocks/financialReporting/trace.mock";
 import { error_issued } from "@mocks/financialReporting/error.mock";
-import { credit_profileInfo } from "@mocks/creditProfileInfo/jobStability.mock";
+import { credit_profileInfo } from "@src/mocks/creditProfileInfo/creditProfileInfo.mock";
 
 export function initializeDataDB() {
   localforage.clear();
